+++ sismic/code/evaluator.py
@@ -1,4 +1,5 @@
 import abc
+from .sequence import Sequence, build_sequence
 from sismic.model import ActionStateMixin
 from sismic.model import Event, Transition, StateMixin, Statechart
 from typing import cast, Iterable, Mapping, List, Dict, Any
@@ -23,7 +24,7 @@ class Evaluator(metaclass=abc.ABCMeta):
     """
     @abc.abstractmethod
     def __init__(self, interpreter=None, *, initial_context: Mapping[str, Any]=None) -> None:
-        pass
+        self._condition_sequences = {}  # type: Dict[str, Dict[str, Sequence]]
 
     @property
     @abc.abstractmethod
@@ -33,6 +34,14 @@ class Evaluator(metaclass=abc.ABCMeta):
         variables and values that is expected to be exposed when the code is evaluated.
         """
         raise NotImplementedError()
+
+    def on_step_starts(self, event: Event=None) -> None:
+        """
+        Called each time the interpreter starts a step.
+
+        :param event: Optional processed event
+        """
+        pass
 
     @abc.abstractmethod
     def _evaluate_code(self, code: str, *, additional_context: Mapping[str, Any]=None) -> bool:
@@ -163,3 +172,48 @@ class Evaluator(metaclass=abc.ABCMeta):
         return filter(
             lambda c: not self._evaluate_code(c, additional_context=event_d), getattr(obj, 'postconditions', [])
         )
+
+    def initialize_sequential_conditions(self, state: StateMixin) -> None:
+        """
+        Initialize sequential conditions.
+
+        :param state: for given state.
+        """
+        condition_mapping = {}  # type: Dict[str, Sequence]
+
+        for condition in getattr(state, 'sequences', []):
+            condition_mapping[condition] = build_sequence(condition, lambda p: self._evaluate_code(p))
+
+        self._condition_sequences[state.name] = condition_mapping
+
+    def update_sequential_conditions(self, state: StateMixin) -> Iterable[str]:
+        """
+        Update sequential conditions, and return a list of already unsatisfied conditions.
+
+        :param state: for given state
+        :return: a list of already unsatisfied conditions.
+        """
+        returned_conditions = []  # type: List[str]
+
+        for condition, sequence in self._condition_sequences[state.name].items():
+            value = sequence.evaluate()
+            if value is False:
+                returned_conditions.append(condition)
+        return returned_conditions
+
+    def evaluate_sequential_conditions(self, state: StateMixin) -> Iterable[str]:
+        """
+        Evaluate sequential conditions, and return a list of unsatisfied conditions.
+
+        :param state: for given state
+        :return: a list of unsatisfied conditions.
+        """
+        returned_conditions = []  # type: List[str]
+
+        for condition, sequence in self._condition_sequences[state.name].items():
+            value = sequence.evaluate(force=True)
+            if value is False:
+                returned_conditions.append(condition)
+        del self._condition_sequences[state.name]
+
+        return returned_conditions